pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.8.3"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
psycopg2 = "^2.8.5"
django-heroku = "^0.3.1"
gunicorn = "^20.0.4"
django-registration-redux = "^2.8"
django-markdownify = "^0.8.0"
djangorestframework = "^3.11.0"
pygments = "^2.6.1"
httpie = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
